Moved components into a directory
