+++ src/web/components/Card/components/ImageLeftLayout.tsx
