+++ src/web/components/Card/components/ImageRightLayout.tsx
